// frontend/vite-project/src/types/DataTypes.ts
export interface DisplayInterface {
    price: number;
    liveFlatPrice: number;
    imgSrc: string;
    liveBroadcastId: number;
    broadcastTitle: string;
}